chore(android): centralize user defined attributes validation
We have multiple track methods in the public API
which can have user defined attributes, however,
the validation is not added in all cases.

This PR centralizes the validation by adding it to
SignalProcessor.
